Complete overhaul
New approach used, making each song a class instance, and storing information in JSON files. This method allows the user to make changes in the console, and for this information to be saved for future use. I also hope eventually to develop a GUI for it, but will have to set aside some time for learning how.