First successful compile
First commit in a while. Most recent advances are in creating a dialogue window to select folder containing the editable JSON files. This is so that these files are separate from the app, and can be stored in a shared folder, so that changes are accessible to all users, although admittedly, this does not allow for simultaneous edits.